apps/server/requirements.txt
Django>=5.1
djangorestframework>=3.15
PyJWT>=2.9
requests>=2.32
dj-database-url>=2.1
python-dotenv>=1.0
django-cors-headers>=4.6
PyMySQL>=1.1
